Proyecto Funcional , con Crud de Usuarios y Estudiantes

+++ surotec-academy/src/main/java/com/website/surotec_academy/repository/StudentRepository.java
@@ -1,15 +1,14 @@
 package com.website.surotec_academy.repository;
 
-import com.website.surotec_academy.classification.StudentStatus;
+import com.website.surotec_academy.enums.StudentStatus;
 import com.website.surotec_academy.entity.StudentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Optional;
 
 @Repository
-public interface StudentRepository extends JpaRepository<StudentEntity, Integer> {
+public interface StudentRepository extends JpaRepository<StudentEntity, Long> {
 
     List<StudentEntity> findByStatus(StudentStatus status);
